+++ Flow/src/utopia/flow/parse/file/FileExtensions.scala
@@ -1,12 +1,15 @@
 package utopia.flow.parse.file
 
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.mutable.iterator.PollableOnce
+import utopia.flow.collection.immutable.caching.LazyTree
+import utopia.flow.collection.mutable.iterator.{OptionsIterator, PollableOnce}
 import utopia.flow.operator.EqualsExtensions._
 import utopia.flow.parse.json.JsonConvertible
 import utopia.flow.parse.string.IterateLines
 import utopia.flow.parse.AutoClose._
 import utopia.flow.util.StringExtensions._
+import utopia.flow.util.logging.Logger
+import utopia.flow.view.immutable.caching.Lazy
 
 import java.awt.Desktop
 import java.io._
@@ -35,7 +38,6 @@ object FileExtensions
 		  * @return Names of the parts that form this path
 		  */
 		def parts = (0 to p.getNameCount).map { p.getName(_).toString }
-		
 		/**
 		  * @return An iterator that returns the names of the parts that form this path
 		  */
@@ -55,7 +57,6 @@ object FileExtensions
 		  * @return Whether this file exists in the file system (false if undetermined)
 		  */
 		def exists = Files.exists(p)
-		
 		/**
 		  * @return Whether this file doesn't exist in the file system (false if undetermined)
 		  */
@@ -65,7 +66,6 @@ object FileExtensions
 		  * @return File name portion of this path
 		  */
 		def fileName = Option(p.getFileName).map { _.toString }.getOrElse("")
-		
 		/**
 		  * @return File name portion of this path, without the extension portion (such as ".txt")
 		  */
@@ -73,18 +73,26 @@ object FileExtensions
 			case Some(part) => part.toString.untilLast(".")
 			case None => ""
 		}
-		
+		/**
+		  * @return The name and the extension of this file
+		  */
+		def fileNameAndType = fileName.splitAtLast(".")
 		/**
 		  * @return The last modified time of this file (may fail). May not work properly for directories.
 		  */
 		def lastModified = Try { Files.getLastModifiedTime(p).toInstant }
-		
 		/**
 		  * @return the type of this file (portion after last '.'). Returns an empty string for directories and
 		  *         for files without type.
 		  */
 		def fileType = fileName.afterLast(".")
 		
+		/**
+		  * @return True if this path is not absolute.
+		  *         Absolute means that this path doesn't need to be combined with another path in order
+		  *         to locate a file.
+		  */
+		def isNotAbsolute = !p.isAbsolute
 		/**
 		  * @return An absolute path based on this path (if this path is already absolute, returns this)
 		  */
@@ -94,40 +102,37 @@ object FileExtensions
 		  * @return Whether this path represents an existing directory
 		  */
 		def isDirectory = Files.isDirectory(p)
-		
 		/**
 		  * @return Whether this path represents an existing regular file (non-directory)
 		  */
 		def isRegularFile = Files.isRegularFile(p)
 		
 		/**
-		  * @return A parent path for this path. None if this path is already a root path
+		  * @return A parent path for this path. None if this path is already a root path.
 		  */
 		def parentOption = Option(p.getParent)
-		
+			.orElse {
+				// Case: Absolute root path => None
+				if (p.isAbsolute)
+					None
+				// Case: Relative path => Converts to an absolute path (if possible) and tries again
+				else
+					Try { p.toAbsolutePath }.toOption.flatMap { p => Option(p.getParent) }
+			}
 		/**
 		  * @return A parent path for this path. Return this path if already a root path
 		  */
-		def parent = parentOption.orElse { p.toAbsolutePath.parentOption }.getOrElse(p)
-		
+		def parent = parentOption.getOrElse(p)
 		/**
 		  * @return An iterator that returns all parents of this path, from closest to furthest
 		  */
-		def parentsIterator = parentOption match {
-			case Some(parent) => Iterator.unfold(parent) { _.parentOption.map { p => p -> p } }
-			case None => Iterator.empty
-		}
+		def parentsIterator = OptionsIterator.iterate(parentOption) { _.parentOption }
 		
 		/**
-		  * @return All children (files and directories) directly under this directory (empty vector if not directory). May fail.
+		  * @return All children (files and directories) directly under this directory
+		  *         (empty vector if not directory). May fail.
 		  */
 		def children = iterateChildren { _.toVector }
-		
-		/**
-		  * @return Directories directly under this one (returns empty vector for regular files). May fail.
-		  */
-		def subDirectories = iterateChildren { _.filter { _.isDirectory }.toVector }
-		
 		/**
 		  * @return An iterator that accesses all child paths within this directory.
 		  *         The iterator may terminate and return failure on read failures.
@@ -136,7 +141,10 @@ object FileExtensions
 			case Success(children) => children.iterator.flatMap { new RecursiveDirectoryIterator(_) }
 			case Failure(error) => PollableOnce(Failure(error))
 		}
-		
+		/**
+		  * @return Directories directly under this one (returns empty vector for regular files). May fail.
+		  */
+		def subDirectories = iterateChildren { _.filter { _.isDirectory }.toVector }
 		/**
 		  * @return An iterator that returns all sub-directories of this directory, their sub-directories and so-on.
 		  *         Includes the regular files with each directory, also.
@@ -146,7 +154,6 @@ object FileExtensions
 				children.iterator.filter { _.isDirectory }.flatMap { new RecursiveDirectoriesIterator(_) }
 			case Failure(error) => PollableOnce(Failure(error))
 		}
-		
 		/**
 		  * @return An iterator that returns this directory, all sub-directories under this directory,
 		  *         all their sub-directories, and so on.
@@ -158,7 +165,6 @@ object FileExtensions
 			else
 				Iterator.empty
 		}
-		
 		/**
 		  * @return All non-directory files in this directory and its sub-directories
 		  */
@@ -168,8 +174,7 @@ object FileExtensions
 		  * @return This path as an existing directory (creating the directory if possible & necessary).
 		  *         Fails if this is a regular file and not a directory.
 		  */
-		def asExistingDirectory =
-		{
+		def asExistingDirectory = {
 			if (notExists)
 				createDirectories()
 			else if (isRegularFile)
@@ -182,8 +187,7 @@ object FileExtensions
 		  * @return This path in case it doesn't exist yet. Otherwise another non-existing path similar name
 		  *         in the same directory.
 		  */
-		def unique =
-		{
+		def unique = {
 			// Case: This path doesn't exist yet => Can't conflict
 			if (notExists)
 				p
@@ -201,6 +205,8 @@ object FileExtensions
 						'_'
 					else if (myName.containsMany("."))
 						'.'
+					else if (myName.contains(' '))
+						' '
 					else
 						'-'
 				}
@@ -218,8 +224,7 @@ object FileExtensions
 		  * @return The size of this file in bytes. If called for a directory, returns the combined size of all files and
 		  *         directories under this directory. Please note that this method may take a while to complete.
 		  */
-		def size: Try[Long] =
-		{
+		def size: Try[Long] = {
 			// Size of a regular file is delegated to java.nio.Files while size of a directory is calculated recursively
 			if (isRegularFile)
 				Try { Files.size(p) }
@@ -247,6 +252,35 @@ object FileExtensions
 			}
 		}
 		
+		/**
+		  * Lazily converts this path (and its sub-directories and files) into a tree structure
+		  * @param log A logger to use to log possible failures to scan for directory children
+		  * @return A lazily initialized tree structure based on this path,
+		  *         where each node represents a file or a directory.
+		  */
+		def toTree(implicit log: Logger): LazyTree[Path] = {
+			// Case: Regular file => Doesn't need any scanning
+			if (isRegularFile)
+				LazyTree.initializedEmpty(p)
+			// Case: Directory => lazily reads the directory contents when going downwards
+			else
+				LazyTree.iterate[Path](Lazy.initialized(p)) { path =>
+					// Case: Directory => lazily scans further
+					if (path.isDirectory)
+						path.children match {
+							// Case: Directory scanning succeeded => opens a new layer of nodes
+							case Success(children) => children.map(Lazy.initialized)
+							// Case: Directory reading failed => logs as an error and acts as if the directory was empty
+							case Failure(error) =>
+								log(error, s"Failed to read the children of $path")
+								Vector.empty
+						}
+					// Case: Regular file => No need to scan for child nodes
+					else
+						Vector.empty
+				}
+		}
+		
 		
 		// OTHER    -------------------------------
 		
@@ -255,7 +289,6 @@ object FileExtensions
 		  * @return This path extended with another path
 		  */
 		def /(another: Path) = p.resolve(another)
-		
 		/**
 		  * @param another A sub-path
 		  * @return This path extended with another path
@@ -268,7 +301,6 @@ object FileExtensions
 		  */
 		def take(count: Int) =
 			if (count >= p.getNameCount) p else if (count <= 0) Paths.get("") else p.subpath(0, count)
-		
 		/**
 		  * @param n Number of elements to drop (from the root)
 		  * @return A copy of this path without the first n parts.
@@ -285,7 +317,6 @@ object FileExtensions
 					p.subpath(n, l)
 			}
 		}
-		
 		/**
 		  * @param count Number of elements to take (from the right)
 		  * @return A copy of this path that's at most 'count' length, with the same last item
@@ -301,7 +332,6 @@ object FileExtensions
 					p.subpath(l - count, l)
 			}
 		}
-		
 		/**
 		  * @param n Number of elements to drop from the right
 		  * @return The nth parent of this path
@@ -326,14 +356,12 @@ object FileExtensions
 		  * @tparam A Type of returned value
 		  * @return The returned value. Failure if something threw during this operation.
 		  */
-		def iterateChildren[A](f: Iterator[Path] => A) =
-		{
+		def iterateChildren[A](f: Iterator[Path] => A) = {
 			if (isDirectory)
 				Try { Files.list(p).consume { stream => f(stream.iterator().asScala) } }
 			else
 				Try { f(Iterator.empty) }
 		}
-		
 		/**
 		  * Iterates over the children of this directory
 		  * @param f A function that accepts an iterator that returns all paths that are the children of this directory.
@@ -368,7 +396,6 @@ object FileExtensions
 			val (common, relative, otherRelative) = commonParentWith(Vector(other))
 			(common, relative, otherRelative.head)
 		}
-		
 		/**
 		  * Seeks the lowest common parent with another path
 		  * @param others Other paths
@@ -390,7 +417,6 @@ object FileExtensions
 				case None => (None, p, others)
 			}
 		}
-		
 		def commonParentWith(other: Path, second: Path, more: Path*): (Option[Path], Path, Seq[Path]) =
 			commonParentWith(Vector(other, second) ++ more)
 		
@@ -432,15 +458,13 @@ object FileExtensions
 		  * @return Whether this directory contains the specified file (false if this is not a directory)
 		  */
 		def containsDirect(childFileName: String) = (this / childFileName).exists
-		
 		/**
 		  * Checks whether this directory or any sub-directory within this directory contains a file with the
 		  * specified name (case-insensitive).
 		  * @param childFileName Name of the searched file (including file extension)
 		  * @return Whether this directory system contains a file with the specified name
 		  */
-		def containsRecursive(childFileName: String): Boolean =
-		{
+		def containsRecursive(childFileName: String): Boolean = {
 			iterateChildren {
 				_.exists { child =>
 					(child.fileName ~== childFileName) || child.containsRecursive(childFileName)
@@ -452,8 +476,7 @@ object FileExtensions
 		  * @param newFileName New file name (may or may not contain an extension)
 		  * @return A copy of this path with specified file name (NB: No file is being renamed as part of this operation)
 		  */
-		def withFileName(newFileName: String) =
-		{
+		def withFileName(newFileName: String) = {
 			val myName = fileName
 			// Case: Already has that file name => returns self
 			if (myName == newFileName)
@@ -471,7 +494,6 @@ object FileExtensions
 				}
 			}
 		}
-		
 		/**
 		  * @param f a file name mapping function
 		  * @return A copy of this path that has the mapped file name
@@ -487,7 +509,6 @@ object FileExtensions
 		@deprecated("Please use the new, more flexible, .iterateChildren(...) instead", "v1.11.2")
 		def forChildren(filter: Path => Boolean = _ => true)(operation: Path => Unit): Try[Unit] =
 			iterateChildren { _.filter(filter).foreach(operation) }
-		
 		/**
 		  * Merges values of child paths into a single value
 		  * @param start  Starting value
@@ -507,7 +528,6 @@ object FileExtensions
 		  */
 		def findDescendants(filter: Path => Boolean): Try[Vector[Path]] =
 			allChildrenIterator.tryFlatMap { _.map { Some(_).filter(filter) } }.map { _.toVector }
-		
 		/**
 		  * @param extension A file extension (Eg. "png"), not including the '.'
 		  * @return All files directly or indirectly under this directory that have the specified file extension / type
@@ -555,7 +575,6 @@ object FileExtensions
 				}
 			}
 		}
-		
 		/**
 		  * Copies this file / directory to another directory
 		  * @param targetDirectory Target parent directory for this file
@@ -564,7 +583,6 @@ object FileExtensions
 		  * @return Link to the target path. Failure if file moving failed or if couldn't replace an existing file
 		  */
 		def copyTo(targetDirectory: Path, replaceIfExists: Boolean = true) = copyAs(targetDirectory / fileName, replaceIfExists)
-		
 		/**
 		  * Copies this file / directory to a new location (over specified path)
 		  * @param targetPath   Location, including file name, for the new copy
@@ -598,9 +616,7 @@ object FileExtensions
 				prepareResult.flatMap { target => recursiveCopyAs(target) }
 			}
 		}
-		
 		private def recursiveCopyTo(targetDirectory: Path): Try[Path] = recursiveCopyAs(targetDirectory / fileName)
-		
 		// First copies the file / directory, then the children files, if there are any
 		private def recursiveCopyAs(newPath: Path): Try[Path] =
 		{
@@ -610,7 +626,6 @@ object FileExtensions
 					.map { _ => newParent }
 			}
 		}
-		
 		/**
 		  * Moves and possibly renames this file
 		  * @param targetPath   New path for this file, including the file name
@@ -619,8 +634,7 @@ object FileExtensions
 		  * @return This file's new path. Failure if moving or file deletion failed or if tried to overwrite a file
 		  *         while allowReplace = false.
 		  */
-		def moveAs(targetPath: Path, allowReplace: Boolean = true) =
-		{
+		def moveAs(targetPath: Path, allowReplace: Boolean = true) = {
 			if (targetPath == p)
 				Success(p)
 			else
@@ -628,7 +642,6 @@ object FileExtensions
 					delete().map { _ => newPath }
 				}
 		}
-		
 		/**
 		  * Renames this file or directory
 		  * @param newFileName    New name for this file or directory (just file name, not the full path)
@@ -636,8 +649,7 @@ object FileExtensions
 		  *                       false, fails when trying to rename over an existing file.
 		  * @return Path to the newly named file. Failure if renaming failed.
 		  */
-		def rename(newFileName: String, allowOverwrite: Boolean = false) =
-		{
+		def rename(newFileName: String, allowOverwrite: Boolean = false) = {
 			// Might not need to rename the file at all
 			if (fileName == newFileName)
 				Success(p)
@@ -659,8 +671,7 @@ object FileExtensions
 		  * @param anotherPath A path leading to the file that will overwrite this one
 		  * @return Path to this file. May contain failure.
 		  */
-		def overwriteWith(anotherPath: Path) =
-		{
+		def overwriteWith(anotherPath: Path) = {
 			// May not need to do anything
 			if (anotherPath == p)
 				Success(p)
@@ -680,7 +691,6 @@ object FileExtensions
 				}
 			}
 		}
-		
 		/**
 		  * Overwrites this path with file from another path, but only if the file was changed (had different last
 		  * modified time). In case where directories are being overwritten, checks each file within the directories
@@ -688,8 +698,7 @@ object FileExtensions
 		  * @param anotherPath Another file that will overwrite this one
 		  * @return Path to this file. May contain a failure
 		  */
-		def overwriteWithIfChanged(anotherPath: Path): Try[Path] =
-		{
+		def overwriteWithIfChanged(anotherPath: Path): Try[Path] = {
 			// May not need to do anything
 			if (anotherPath == p)
 				Success(p)
@@ -761,7 +770,6 @@ object FileExtensions
 					}
 				}
 		}
-		
 		/**
 		  * Deletes all files from under this directory
 		  * @return Success containing this directory if all deletions succeeded.
@@ -770,7 +778,6 @@ object FileExtensions
 		def deleteContents() =
 			iterateChildren { _.map { _.delete() }.toVector.find { _.isFailure }.getOrElse { Success(()) } }
 				.flatten.map { _ => p }
-		
 		/**
 		  * Deletes all child paths from under this directory. Stops deletion if any deletion fails.
 		  * @return Whether any files were deleted. May contain failure.
@@ -783,8 +790,7 @@ object FileExtensions
 		  * simply creates parent directories.
 		  * @return This path. Failure if couldn't create directories.
 		  */
-		def createDirectories() =
-		{
+		def createDirectories() = {
 			if (notExists) {
 				// Checks whether this file should be a directory (doesn't have a file type) or a regular file
 				// (has file type)
@@ -796,20 +802,18 @@ object FileExtensions
 			else
 				Success(p)
 		}
-		
 		/**
 		  * Creates directories above this path. Eg. for path "dir1/dir2/fileX.txt" would ensure existence of dir1 and dir2
 		  * @return This path, failure if couldn't create directories
 		  */
-		def createParentDirectories() = parentOption.map { dir => Try[Unit] { Files.createDirectories(dir) } }
-			.getOrElse(Success(())).map { _ => p }
+		def createParentDirectories() =
+			parentOption.map { dir => Try[Unit] { Files.createDirectories(dir) } }.getOrElse(Success(())).map { _ => p }
 		
 		/**
 		  * @param another Another file
 		  * @return Whether these two files have same last modified time
 		  */
-		def hasSameLastModifiedAs(another: Path): Boolean =
-		{
+		def hasSameLastModifiedAs(another: Path): Boolean = {
 			// Directories need to be handled a bit differently (files inside the directory may have changed)
 			val directLastModifiedComparison = lastModified.success.exists { another.lastModified.success.contains }
 			if (isDirectory && another.isDirectory) {
@@ -833,7 +837,6 @@ object FileExtensions
 		  * @return This path. Failure if writing failed.
 		  */
 		def write(text: String)(implicit codec: Codec) = Try { Files.write(p, text.getBytes(codec.charSet)) }
-		
 		/**
 		  * Writes a json-convertible instance to this file
 		  * @param json A json-convertible instance that will produce contents of this file
@@ -841,14 +844,12 @@ object FileExtensions
 		  */
 		@deprecated("Replaced with writeJson", "v1.9")
 		def writeJSON(json: JsonConvertible) = write(json.toJson)(Codec.UTF8)
-		
 		/**
 		  * Writes a json-convertible instance to this file
 		  * @param json A json-convertible instance that will produce contents of this file
 		  * @return This path. Failure if writing failed.
 		  */
 		def writeJson(json: JsonConvertible) = write(json.toJson)(Codec.UTF8)
-		
 		/**
 		  * Writes the specified text lines to a file
 		  * @param lines  Lines to write to the file
@@ -856,8 +857,7 @@ object FileExtensions
 		  *               to overwrite the current contents of the file (false). Default = false.
 		  * @param codec  Encoding used (implicit)
 		  */
-		def writeLines(lines: IterableOnce[String], append: Boolean = false)(implicit codec: Codec) =
-		{
+		def writeLines(lines: IterableOnce[String], append: Boolean = false)(implicit codec: Codec) = {
 			Try {
 				new FileOutputStream(p.toFile, append)
 					.consume {
@@ -870,14 +870,12 @@ object FileExtensions
 					}
 			}
 		}
-		
 		/**
 		  * Writes into this file with a function. An output stream is opened for the duration of the function.
 		  * @param writer A writer function that uses an output stream (may throw)
 		  * @return This path. Failure if writing function threw or stream couldn't be opened
 		  */
 		def writeWith[U](writer: BufferedOutputStream => U) = _writeWith(append = false)(writer)
-		
 		/**
 		  * Writes a file using a function.
 		  * A PrintWriter instance is acquired for the duration of the function execution.
@@ -888,16 +886,13 @@ object FileExtensions
 		  */
 		def writeUsing[U](writer: PrintWriter => U)(implicit codec: Codec) =
 			_writeUsing(append = false)(writer)
-		
 		def _writeUsing[U](append: Boolean)(writer: PrintWriter => U)(implicit codec: Codec) =
 			_writeWith(append) { stream =>
 				stream.consume { new OutputStreamWriter(_, codec.charSet).consume { new PrintWriter(_).consume(writer) } }
 			}
-		
 		private def _writeWith[U](append: Boolean)(writer: BufferedOutputStream => U) =
 			Try { new FileOutputStream(p.toFile, append).consume { new BufferedOutputStream(_).consume(writer) } }
 				.map { _ => p }
-		
 		/**
 		  * Writes into this file by reading data from a reader.
 		  * @param reader Reader that supplies the data
@@ -912,7 +907,6 @@ object FileExtensions
 				.takeWhile { _ != -1 }
 				.foreach(output.write)
 		}
-		
 		/**
 		  * Writes the specified input stream into this file
 		  * @param inputStream Reader that supplies the data
@@ -922,7 +916,6 @@ object FileExtensions
 			Files.copy(inputStream, p,
 				StandardCopyOption.REPLACE_EXISTING)
 		}.map { _ => p }
-		
 		/**
 		  * Appends specified text to this file
 		  * @param text  Text to append to this file
@@ -933,7 +926,6 @@ object FileExtensions
 			Files.write(p, text.getBytes(codec.charSet),
 				StandardOpenOption.APPEND, StandardOpenOption.CREATE)
 		}
-		
 		/**
 		  * Writes into this file with a function. An output stream is opened for the duration of the function.
 		  * Doesn't overwrite the current contents of this file but appends to them instead.
@@ -941,7 +933,6 @@ object FileExtensions
 		  * @return This path. Failure if writing function threw or stream couldn't be opened
 		  */
 		def appendWith[U](writer: BufferedOutputStream => U) = _writeWith(append = true)(writer)
-		
 		/**
 		  * Appends new lines to a file utilizing a PrintWriter
 		  * @param writer A function that uses a PrintWriter
@@ -951,14 +942,12 @@ object FileExtensions
 		  */
 		def appendUsing[U](writer: PrintWriter => U)(implicit codec: Codec) =
 			_writeUsing(append = true)(writer)
-		
 		/**
 		  * Writes the specified text lines to the end of this file
 		  * @param lines Lines to write to the file
 		  * @param codec Encoding used (implicit)
 		  */
 		def appendLines(lines: IterableOnce[String])(implicit codec: Codec) = writeLines(lines, append = true)
-		
 		/**
 		  * Writes into this file by reading data from a reader. Doesn't overwrite existing file data but appends
 		  * to it instead.
@@ -974,7 +963,6 @@ object FileExtensions
 		  * @return Returned value or failure if stream couldn't be opened / read or the reader function threw.
 		  */
 		def readWith[A](reader: FileInputStream => A) = Try { new FileInputStream(p.toFile).consume(reader) }
-		
 		/**
 		  * Reads data from this file
 		  * @param reader A function that reads this file's data stream and returns a Try
@@ -992,8 +980,7 @@ object FileExtensions
 		  * @return Edited copy path. Failure if this path was a directory or didn't exist,
 		  *         or if the writing or reading failed.
 		  */
-		def editToCopy[U](copyPath: Path)(f: FileEditor => U)(implicit codec: Codec) =
-		{
+		def editToCopy[U](copyPath: Path)(f: FileEditor => U)(implicit codec: Codec) = {
 			// Makes sure this is an existing regular file
 			if (isDirectory)
 				Failure(new IOException("Directories can't be edited using .editToCopy(...)"))
@@ -1010,7 +997,6 @@ object FileExtensions
 					}
 				}
 		}
-		
 		/**
 		  * Edits this file, saving the edited copy as a separate file in the same directory
 		  * @param copyName Name of the edited copy file (extension isn't required)
@@ -1022,7 +1008,6 @@ object FileExtensions
 		  */
 		def editToCopy[U](copyName: String)(f: FileEditor => U)(implicit codec: Codec): Try[Path] =
 			editToCopy(withFileName(copyName))(f)
-		
 		/**
 		  * Edits the contents of this file. The edits actualize at the end of this method call.
 		  * @param f     A function that uses a file editor to make the edits
@@ -1031,8 +1016,7 @@ object FileExtensions
 		  * @return This path. Failure if this file was not editable (e.g. non-existing or a directory) or if
 		  *         reading, writing, or replacing failed
 		  */
-		def edit[U](f: FileEditor => U)(implicit codec: Codec) =
-		{
+		def edit[U](f: FileEditor => U)(implicit codec: Codec) = {
 			// Finds a copy name that hasn't been taken yet
 			val (fileNamePart, extensionPart) = fileName.splitAtLast(".")
 			// Writes to copy by editing the original
@@ -1048,20 +1032,17 @@ object FileExtensions
 		  * @return Success of failure
 		  */
 		def openInDesktop() = performDesktopOperation { _.open(p.toFile) }
-		
 		/**
 		  * Opens this file for editing in the default desktop application. If this is a directory, opens it in resource manager
 		  * @return Success or failure
 		  */
 		def editInDesktop() =
 			performDesktopOperation { d => if (isDirectory) d.open(p.toFile) else d.edit(p.toFile) }
-		
 		/**
 		  * Prints this file using the default desktop application
 		  * @return Success or failure
 		  */
 		def print() = performDesktopOperation { _.print(p.toFile) }
-		
 		/**
 		  * Opens resource manager for this directory, or the directory containing this file
 		  * @return Success or failure
@@ -1069,15 +1050,13 @@ object FileExtensions
 		def openDirectory() = performDesktopOperation { d =>
 			if (isDirectory) d.open(p.toFile) else d.open(parent.toFile)
 		}
-		
 		/**
 		  * Opens resource manager for this file's / directory's parent directory
 		  * @return Success or failure
 		  */
 		def openFileLocation() = performDesktopOperation { _.open(parent.toFile) }
 		
-		private def performDesktopOperation(f: Desktop => Unit) =
-		{
+		private def performDesktopOperation(f: Desktop => Unit) = {
 			if (Desktop.isDesktopSupported)
 				Try { f(Desktop.getDesktop) }
 			else